refactor: retrieve genesis file url from chain
