+++ utils/config/tomlconfig/toml.go
@@ -1,13 +1,22 @@
 package tomlconfig
 
 import (
+	"crypto/sha256"
+	"encoding/hex"
 	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 
+	comettypes "github.com/cometbft/cometbft/types"
 	naoinatoml "github.com/naoina/toml"
+	"github.com/pterm/pterm"
 
+	initconfig "github.com/dymensionxyz/roller/cmd/config/init"
 	"github.com/dymensionxyz/roller/cmd/consts"
 	"github.com/dymensionxyz/roller/utils/bash"
 	"github.com/dymensionxyz/roller/utils/config"
@@ -71,6 +80,8 @@ func LoadRollappMetadataFromChain(
 		cfg = config.RollappConfig{
 			Home:             home,
 			RollappID:        "mock_1000-1",
+			GenesisHash:      "",
+			GenesisUrl:       "",
 			RollappBinary:    consts.Executables.RollappEVM,
 			VMType:           consts.EVM_ROLLAPP,
 			Denom:            "mock",
@@ -107,6 +118,8 @@ func LoadRollappMetadataFromChain(
 
 		cfg = config.RollappConfig{
 			Home:             home,
+			GenesisHash:      raResponse.Rollapp.GenesisChecksum,
+			GenesisUrl:       raResponse.Rollapp.Metadata.GenesisUrl,
 			RollappID:        raResponse.Rollapp.RollappId,
 			RollappBinary:    consts.Executables.RollappEVM,
 			VMType:           consts.EVM_ROLLAPP,
@@ -121,7 +134,102 @@ func LoadRollappMetadataFromChain(
 			BaseDenom:        "amock",
 			MinGasPrices:     "0",
 		}
+
+		genesisPath := initconfig.GetGenesisFilePath(home)
+		genesisUrl := raResponse.Rollapp.Metadata.GenesisUrl
+		err = downloadFile(genesisUrl, genesisPath)
+		if err != nil {
+			return nil, err
+		}
+
+		// move to helper function with a spinner?
+		genesis, err := comettypes.GenesisDocFromFile(genesisPath)
+		if err != nil {
+			return nil, err
+		}
+
+		if genesis.ChainID != raID {
+			err = fmt.Errorf(
+				"the genesis file ChainID (%s) does not match  the rollapp ID you're trying to initialize ("+
+					"%s)",
+				genesis.ChainID,
+				raID,
+			)
+			return nil, err
+		}
+
+		downloadedGenesisHash, err := calculateSHA256(genesisPath)
+		if err != nil {
+			pterm.Error.Println("failed to calculate hash of genesis file: ", err)
+		}
+		raGenesisHash, _ := getRollappGenesisHash(raID, *hd)
+		if downloadedGenesisHash != raGenesisHash {
+			err = errors.New(
+				"the hash of the downloaded file does not match the one registered with the rollapp",
+			)
+			return nil, err
+		}
 	}
 
 	return &cfg, nil
+}
+
+// TODO: download the file in chunks if possible
+func downloadFile(url, filepath string) error {
+	spinner, _ := pterm.DefaultSpinner.
+		Start("Downloading genesis file from ", url)
+
+	// nolint:gosec
+	resp, err := http.Get(url)
+	if err != nil {
+		spinner.Fail("failed to download file: ", err)
+		return err
+	}
+	defer resp.Body.Close()
+
+	out, err := os.Create(filepath)
+	if err != nil {
+		spinner.Fail("failed to download file: ", err)
+		return err
+	}
+	defer out.Close()
+
+	spinner.Success("Successfully downloaded the genesis file")
+	_, err = io.Copy(out, resp.Body)
+	return err
+}
+
+func calculateSHA256(filepath string) (string, error) {
+	file, err := os.Open(filepath)
+	if err != nil {
+		return "", fmt.Errorf("error opening file: %v", err)
+	}
+	defer file.Close()
+
+	hash := sha256.New()
+	if _, err := io.Copy(hash, file); err != nil {
+		return "", fmt.Errorf("error calculating hash: %v", err)
+	}
+
+	return hex.EncodeToString(hash.Sum(nil)), nil
+}
+
+func getRollappGenesisHash(raID string, hd consts.HubData) (string, error) {
+	var raResponse rollapp.ShowRollappResponse
+	getRollappCmd := exec.Command(
+		consts.Executables.Dymension,
+		"q", "rollapp", "show",
+		raID, "-o", "json", "--node", hd.RPC_URL,
+	)
+
+	out, err := bash.ExecCommandWithStdout(getRollappCmd)
+	if err != nil {
+		return "", err
+	}
+
+	err = json.Unmarshal(out.Bytes(), &raResponse)
+	if err != nil {
+		return "", err
+	}
+	return raResponse.Rollapp.GenesisChecksum, nil
 }
